cleaning code
At this stage, everything is working as intended, only some parameters need to be adjusted. For the A star and RRT, we need to increase the value of repulsion such that the UAVs don't touch each other, and we'll need to fix the potential field approach where they don't stop spinning around the cone.